chore(parser): move parseKnativeIngressRules()
Move parseKnativeIngressRules() to translate.go (including the test),
and rename to fromKnativeIngress().
